fix: 우승자 구하는 로직 view 에서 domain 로직 이동
